Moving stuff back to get - see notes
Making everything post screws up the browser back capability.
and i don't think it added anything to improve anti-scraping.
But i did leave everything in a form and left the PRG pattern
in place (except is it a "GRG" pattern) and it is easy to switch
from get to post.